+++ src/main.tsx
@@ -1,13 +1,16 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
+import { TooltipProvider } from './components/ui/tooltip.tsx'
 import App from './App.tsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <BrowserRouter>
-      <App />
+      <TooltipProvider delayDuration={200}>
+        <App />
+      </TooltipProvider>
     </BrowserRouter>
   </React.StrictMode>,
 )